XML Methods transfered to XmlHandler.cs Added Image_id XmlElement Updated question loading with images
