+++ Exam/Forms/Teacher/QuestionPage_Teacher.cs
@@ -15,6 +15,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Xml.Linq;
+using DAL;
 
 namespace Exam.Forms.Student
 {
@@ -133,87 +134,10 @@ namespace Exam.Forms.Student
 		#region XML Methods
 		public void CreateXMLFiles()
 		{
-			CreateGradesFile();
-			CreateQuestionsFile();
-		}
-		public void CreateGradesFile()
-		{
-			string titleNoSpace = exam.Title;
-			foreach (var c in titleNoSpace)
-			{
-				titleNoSpace = titleNoSpace.Replace(" ", string.Empty);
-			}
-			string fileName = ($"{titleNoSpace}_{exam.Id}.xml");
-			string dirName = "ExamsGrades";
-			var filePath = SetFilePath(fileName, dirName);
-
-			XElement xEntity = new XElement($"{titleNoSpace}_{exam.Id}");
-			xEntity.Save(filePath);
-			exam.GradesPath = $"{dirName}\\{fileName}";
+			XmlHandler.CreateGradesFile(exam);
+			XmlHandler.CreateQuestionsFile(exam);
 		}
 
-		public void CreateQuestionsFile()
-		{
-			string titleNoSpace = exam.Title;
-			foreach (var c in titleNoSpace)
-			{
-				titleNoSpace = titleNoSpace.Replace(" ", string.Empty);
-			}
-			XElement xEntity = new XElement($"{titleNoSpace}_{exam.Id}");
-			XElement xQuestions = new XElement("Questions");
-			foreach (var question in exam.Questions)
-			{
-				if (question is MultipleChoiceTextQuestion)
-				{
-					MultipleChoiceTextQuestion multipleText = question as MultipleChoiceTextQuestion;
-					XElement xAnswers = new XElement("Answers");
-					int i = 0;
-					foreach (string answer in multipleText.Answers)
-					{
-						xAnswers.Add(new XElement($"Answer_{i}", answer));
-						i++;
-					}
-					XElement xQuestion = new XElement("MultipleChoiseTextQuestion", new object[]{
-					   new XElement("QuestionDescription",multipleText.QuestionDescription),
-					   new XElement("QuestionText",multipleText.QuestionText),
-					   xAnswers,
-					   new XElement("RightAnswer",multipleText.RightAnswer),
-					   new XElement("Points",multipleText.Points)
-					});
-					xQuestions.Add(xQuestion);
-				}
-				else if (question is OpenQuestion)
-				{
-					OpenQuestion openQuestion = question as OpenQuestion;
-					XElement xQuestion = new XElement("OpenQuestion", new object[]{
-					   new XElement("QuestionDescription",openQuestion.QuestionDescription),
-					   new XElement("QuestionText",openQuestion.QuestionText),
-					   new XElement("RightAnswer",openQuestion.RightAnswer),
-					   new XElement("Points",openQuestion.Points)
-					});
-					xQuestions.Add(xQuestion);
-				}
-			}
-			xEntity.Add(xQuestions);
-			//var dirPath = Directory.GetCurrentDirectory();
-			string fileName = ($"{titleNoSpace}_{exam.Id}.xml");
-			string dirName = "ExamsQuestions";
-			var filePath = SetFilePath(fileName, dirName);
-			xEntity.Save(filePath);
-			exam.QuestionsPath = $"{dirName}\\{fileName}";
-		}
-
-
-		private string SetFilePath(string fileName, string dirName)
-		{
-			var dirPath = Directory.GetCurrentDirectory();
-			if (!Directory.Exists($"{dirPath}\\{dirName}"))
-			{
-				Directory.CreateDirectory($"{dirPath}\\{dirName}");
-			}
-			var filePath = $"{dirPath}\\{dirName}\\{fileName}";
-			return filePath;
-		}
 		#endregion
 
 		#endregion
@@ -337,15 +261,15 @@ namespace Exam.Forms.Student
 		}
 
 		// from https://stackoverflow.com/questions/25400555/save-and-retrieve-image-binary-from-sql-server-using-entity-framework-6
-		public Image ByteArrayToImage(byte[] byteArrayIn)
-		{
-			using (var ms = new MemoryStream(byteArrayIn))
-			{
-				var returnImage = Image.FromStream(ms);
+		//public Image ByteArrayToImage(byte[] byteArrayIn)
+		//{
+		//	using (var ms = new MemoryStream(byteArrayIn))
+		//	{
+		//		var returnImage = Image.FromStream(ms);
 
-				return returnImage;
-			}
-		}
+		//		return returnImage;
+		//	}
+		//}
 		#endregion
 
 		public void FormShowDialog()
